+++ internal/tss/utils/transaction_service_test.go
@@ -1,4 +1,4 @@
-package tss_services
+package utils
 
 import (
 	"bytes"
